documentation and class refactoring
