+++ pom.xml
@@ -57,11 +57,11 @@
 
     <modules>
         <module>model</module>
+        <module>persistence-mysql</module>
         <module>runtime</module>
         <module>task</module>
         <module>client</module>
         <module>examples</module>
-        <module>persistence-mysql</module>
     </modules>
 
 </project>
