+++ runtime/src/main/java/com/flipkart/flux/dao/iface/EventsDAO.java
@@ -18,13 +18,17 @@ import com.flipkart.flux.domain.Event;
 import java.util.List;
 
 /**
+ * Provides methods to perform CR operations on {@link com.flipkart.flux.domain.Event}
  * @author shyam.akirala
  */
 public interface EventsDAO {
 
+    /** Creates Event in the db, and returns the saved object*/
     Event create(Event event);
 
+    /** Retrieves all the events which belongs to a particular state machine instance*/
     List<Event> findBySMInstanceId(String stateMachineInstanceId);
 
+    /** Retrieves Event by it's unique identifier*/
     Event findById(Long id);
 }
