+++ runtime/src/main/resources/hibernate.cfg.xml
@@ -10,10 +10,11 @@
         <property name="hibernate.connection.username">root</property>
         <property name="hibernate.current_session_context_class">managed</property>
 
+        <!-- c3po connection pool configuration-->
         <property name="connection.provider_class">org.hibernate.c3p0.internal.C3P0ConnectionProvider</property>
         <property name="hibernate.c3p0.min_size">10</property>
-        <property name="hibernate.c3p0.max_size">50</property>
-        <property name="hibernate.c3p0.max_statements">50</property>
+        <property name="hibernate.c3p0.max_size">100</property>
+
         <mapping package="com.flipkart.flux"/>
     </session-factory>
 </hibernate-configuration>
